whisper-service/requirements.txt
openai-whisper>=20231117
flask>=3.0.0
